% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampliQ.R
\name{ampliQ}
\alias{ampliQ}
\title{A function to graph (ggplot) amplification values (deltaRN vs CT)}
\usage{
ampliQ(df, color, facets)
}
\arguments{
\item{df}{A data frame.}

\item{color}{Variable to dictate color of the graph}

\item{facets}{Include variables for facetwrap; tie two variables with ~ in quotations}
}
\value{
ggplot of Amplification plot
}
\description{
A function to graph (ggplot) amplification values (deltaRN vs CT)
}
\keyword{RN,}
\keyword{amplification,}
\keyword{deltaRN}
