% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amportQ.R
\name{amportQ}
\alias{amportQ}
\title{A function to import amplification data from .xls files.}
\usage{
amportQ(file, sep_var)
}
\arguments{
\item{file}{The ".xls/x" present in working directory or specific ".xls/x".}

\item{sep_var}{Variables to separate the 'Sample' column into.}
}
\value{
Data frame of amplification data.
}
\description{
A function to import amplification data from .xls files.
}
\keyword{amplification}
\keyword{amplification,}
\keyword{curve}
\keyword{import,}
