% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgQ.R
\name{avgQ}
\alias{avgQ}
\title{A function to average CT values by groups}
\usage{
avgQ(df, ..., name = "CT", param_var, ignoreNAs = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Variables to group by.}

\item{name}{Text to append to N and mean arguments}

\item{param_var}{Parameter or gene to be summarize}

\item{ignoreNAs}{Remove NAs, defaults to TRUE}
}
\value{
Data frame of average CT values by groups and number of values averaged.
}
\description{
A function to average CT values by groups
}
\keyword{average}
\keyword{mean,}
\keyword{summarize,}
