% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddQ.R
\name{ddQ}
\alias{ddQ}
\title{A function to calculate delta delta CT values by groups}
\usage{
ddQ(df, ..., col, control, na.rm = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Variables to group by.}

\item{col}{Column containing group of interest.}

\item{control}{String value for control group of interest.}

\item{na.rm}{Remove NAs from data, defaults to TRUE.}
}
\value{
Data frame of average CT values by groups and number of values averaged.
}
\description{
A function to calculate delta delta CT values by groups
}
\keyword{average}
\keyword{mean,}
\keyword{summarize,}
