% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaQ.R
\name{deltaQ}
\alias{deltaQ}
\title{A function to calculate delta CT values by groups}
\usage{
deltaQ(df, ..., ref_gene)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Variables to group by.}

\item{ref_gene}{Reference or housekeeping gene for comparison.}
}
\value{
Data frame of delta CT values by groups.
}
\description{
A function to calculate delta CT values by groups
}
\keyword{delta,}
\keyword{deltaCT}
