% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importQ.R
\name{importQ}
\alias{importQ}
\title{A function to import the 'Results' sheet from .xls/x files into R.}
\usage{
importQ(dir = "")
}
\arguments{
\item{dir}{The specific folder path containing all .xls files, defaults to working directory.}
}
\value{
Combined dataframe containing data from all of the ranges from 'Results' sheets.
}
\description{
A function to import the 'Results' sheet from .xls/x files into R.
}
\keyword{file,}
\keyword{import}
\keyword{qPCR}
\keyword{raw}
