% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importQ.R
\name{importQ}
\alias{importQ}
\title{A function to import the 'Results' sheet from .xls/x files into R.}
\usage{
importQ(file = here(list.files(pattern = "*.xls|*.xlsx")), num_samples)
}
\arguments{
\item{file}{The `.xls/x` present in working directory or specific `.xls/x`.}

\item{num_samples}{The specific range of cells within each sheet in your `.xls/x` file.}
}
\value{
Combined dataframe containing data from all of the ranges from 'Results' sheets.
}
\description{
A function to import the 'Results' sheet from .xls/x files into R.
}
\keyword{file,}
\keyword{import}
\keyword{qPCR}
\keyword{raw}
