% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importQ.R
\name{importQ}
\alias{importQ}
\title{A function to import .xls files into R}
\usage{
importQ(file, ranges)
}
\arguments{
\item{file}{The `.xls` sheet that you want to import (it should be in your working directory).}

\item{ranges}{The specific range of cells within each sheet in your `.xls` file.}
}
\value{
Combined dataframe containing data from all of the ranges from sheets.
}
\description{
A function to import .xls files into R
}
\keyword{file,}
\keyword{import}
\keyword{qPCR}
\keyword{raw}
